moved the utils.py file from /translate to /utils and changed the name to scoring_utils.py
