+++ onmt/trainer.py
@@ -15,7 +15,7 @@ import traceback
 import onmt.utils
 from onmt.utils.loss import LossCompute
 from onmt.utils.logging import logger
-from onmt.translate.utils import ScoringPreparator
+from onmt.utils.scoring_utils import ScoringPreparator
 from onmt.scorers import get_scorers_cls, build_scorers
